refactor: email-detail 컴포넌트 분리 (#40)
